Fix delete session hook to refetch get sessions and to consider local storage

+++ src/frontend/src/controllers/API/queries/messages/use-delete-sessions.ts
@@ -1,26 +1,46 @@
-import type { UseMutationResult } from "@tanstack/react-query";
 import type { useMutationFunctionType } from "@/types/api";
+import type { UseMutationResult } from "@tanstack/react-query";
 import { api } from "../../api";
 import { getURL } from "../../helpers/constants";
 import { UseRequestProcessor } from "../../services/request-processor";
+
+interface UseDeleteSessionParams {
+  flowId?: string;
+  useLocalStorage?: boolean;
+}
 
 interface DeleteSessionParams {
   sessionId: string;
 }
 
 export const useDeleteSession: useMutationFunctionType<
-  undefined,
+  UseDeleteSessionParams,
   DeleteSessionParams
-> = (options?) => {
+> = ({ flowId, useLocalStorage }, options?) => {
   const { mutate, queryClient } = UseRequestProcessor();
 
   const deleteSession = async ({
     sessionId,
   }: DeleteSessionParams): Promise<any> => {
-    const response = await api.delete(
-      `${getURL("MESSAGES")}/session/${sessionId}`,
-    );
-    return response.data;
+    if (!flowId) {
+      throw new Error("Flow ID is required");
+    }
+
+    if (useLocalStorage) {
+      const messages = JSON.parse(sessionStorage.getItem(flowId) || "");
+      const filteredMessages = messages.filter(
+        (message: any) => message.session_id !== sessionId
+      );
+      sessionStorage.setItem(flowId, JSON.stringify(filteredMessages));
+      return {
+        data: filteredMessages,
+      };
+    } else {
+      const response = await api.delete(
+        `${getURL("MESSAGES")}/session/${sessionId}`
+      );
+      return response.data;
+    }
   };
 
   const mutation: UseMutationResult<
@@ -31,7 +51,7 @@ export const useDeleteSession: useMutationFunctionType<
     ...options,
     onSettled: (data, error, variables, context) => {
       queryClient.invalidateQueries({
-        queryKey: ["useGetSessionsFromFlowQuery"],
+        queryKey: ["useGetSessionsFromFlowQuery", { flowId }],
       });
       options?.onSettled?.(data, error, variables, context);
     },
